refactor(analytics): finish shared database imports
