+++ agent/dashboard/routes.py
@@ -1968,7 +1968,7 @@ async def api_agent_usage_leaderboard(
     from asyncpg import PostgresError
     from sqlalchemy.exc import SQLAlchemyError
 
-    from agent.analytics.database import configured
+    from agent.database.analytics import configured
 
     try:
         if not configured():
@@ -1997,7 +1997,7 @@ async def api_pr_merge_rate_by_model(
     from asyncpg import PostgresError
     from sqlalchemy.exc import SQLAlchemyError
 
-    from agent.analytics.database import configured
+    from agent.database.analytics import configured
     from agent.analytics.queries import pr_merge_rate_by_model
 
     try:
@@ -2020,7 +2020,7 @@ async def api_pr_merge_rate_by_model(
 async def api_analytics_readiness(
     _admin: dict[str, Any] = _ADMIN_DEP,
 ) -> dict[str, Any]:
-    from agent.analytics.database import readiness
+    from agent.database.analytics import readiness
 
     return await readiness()
